backend/requirements.txt
fastapi==0.108.0
uvicorn[standard]==0.24.0
psycopg2-binary==2.9.10
python-multipart==0.0.6
requests==2.31.0 
asyncpg==0.30.0
python-dotenv==1.0.0
# 电子围栏功能新增依赖
shapely==2.0.3
pyproj==3.6.1
redis==5.0.1
pydantic==2.10.4
aiofiles==23.2.0